Time: 0 ms (100.00%) | Memory: 8 MB (73.93%) - LeetSync

+++ 50-powx-n/powx-n.cpp
@@ -3,11 +3,8 @@ public:
     double myPow(double x, int n) {
         double ans = 1.0 ;
         long long nn = n  ; // This is done to avoid stack overflow ;
-
         if (nn < 0 ) nn = nn * -1 ;
-
         while (nn)  {
-
             if (nn %2  == 1){
                 ans = ans * x ;
                 nn-- ;
@@ -18,6 +15,6 @@ public:
             }
         }
         if (n < 0) return (double)(1.0)/(double)(ans) ;
-        return (double)(ans) ;
+        return (ans) ;
     }
 };
